pyproject.toml
[project]
name = "ironagent"
version = "0.1.0"
description = "Agentic AI training coach powered by Hevy workout data"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "openai>=1.30",
    "langgraph>=0.2",
    "langchain-openai>=0.2",
    "langchain-core>=0.3",
    "chromadb>=0.5",
    "deepeval>=1.0",
    "ragas>=0.2",
    "fastapi>=0.115",
    "uvicorn>=0.30",
    "streamlit>=1.38",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.5"]